+++ app/javascript/bundles/Coupons/components/LoadingSpinner.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Spinner from 'react-bootstrap/Spinner';
+
+const LoadingSpinner = () => {
+  return (
+    <div className="d-flex flex-lg-column justify-content-center align-items-center mt-5" id="spinner">
+      <Spinner animation="border" role="status" variant="warning">
+        <span className="visually-hidden">Loading...</span>
+      </Spinner>
+    </div>
+  );
+};
+
+export default LoadingSpinner;
