+++ app/views/coupons/index.html.slim
@@ -1,1 +1,1 @@
-= react_component("CouponsIndexPage", props: { path: user_coupons_path }, prerender: false)
+= react_component("CouponsPage", props: { path: user_coupons_path }, prerender: false)
